server/pvlib_api/requirements-datasheet.txt
# Datasheet Processor Requirements
# PDF processing and database integration for solar panel datasheets

# PDF Processing
pdfplumber>=0.9.0

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.27.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Data Validation
pydantic>=2.0.0

# Utilities
tqdm>=4.64.0
python-dotenv>=1.0.0

# Logging
structlog>=22.0.0

# Date/Time
python-dateutil>=2.8.0

# Optional: For better table extraction
camelot-py[cv]>=0.10.0
# ghostscript for camelot (may require system installation)

# Optional: OCR for scanned PDFs
pytesseract>=0.3.10
Pillow>=9.0.0
